refactor mine sweaper out of page into own component. also added missing MuseoModerno font
